UPSTREAM: <carry>: OCPBUGS-18641: Fix dual-stack node addresses in instance metadata
We have discovered that in dual-stack setups `NodeAddresses` field of
the instance metadata contains only IPv4 addresses for VMs that do have
both IPv4 and IPv6 addresses assigned (and detected by the VM agent).

It has been traced back to the function responsible for populating this
metadata field. We found out that for our configuration we always filter
only IPv4 addresses even if running in dual-stack. Reason for this is
the use of the `vcInstance` variable. This variable is a property of
vCenter and not of the VM and as such it returns "ipv4" instead of
"ipv4,ipv6" when needed.

This is later used to cross-check IP addresses provided by the kubelet
as part of the `alpha.kubernetes.io/provided-node-ip` annotation. This
results in removing kubelet's IPv6 addresses and having the final Node
CR contain only IPv4 address as its `Status.Addresses.InternalIP` what
is wrong for dual-stack setups.

We have found that it is safe to forcefully discover IPv4 and IPv6
addresses for all the scenarios (and not only dual-stack) because of how
`nil` values are handled when calling `AddToNodeAddresses` funtion.

Fixes: OCPBUGS-18641
Contributes-to: OPNET-230