usptream: 1532: Make sure to use a non-existent token in tests
When the "should fail when no token is passed and SA token cannot be
read" test is run within a pod on a Kubernetes cluster, the test will
fail because the service account token file is present. This change
provides a path to a file that should not conflict with any others on
the system.

Upstream PR: https://github.com/kubernetes/cloud-provider-vsphere/pull/1532

Signed-off-by: Nolan Brubaker <nolan@nbrubaker.com>